Added two new files, 'general_functions.py' and 'mech_class.py' to the project. These files help with organization.
The 'general_functions.py' file is used to house the functions BeautifulSoup custom functions I created and the "general" custom functions I created to help with strings and dictionaries.

The 'mech_class.py' file is for the class objects "CreatePost" and "QuickCreatePost". "CreatePost" object now holds all the methods used to navigate through the Craigslist site and enter the user's inputs. "QuickCreatePost" inherits the methods of "CreatePost" and allows for user to quickly go through the steps of 'creating a post' using the built-in values to be inputted (using the for sale by owner video games type of posting).

Updated 'mech_functions.py' to only include functions that are used exclusively in the "CreatePost" methods within 'mech_class.py'.
Updated 'fields.py' to reflect the changes made in the other files.