Moved the classes and functions in 'fields.py' to the 'mech_functions.py' file. These classes and functions are used to group fieldset tags during "Step 3 - Add Details To post". Removed the 'fields.py' file.
Updated 'clbot.py' with new updated structure of the project. This file is used to house the main function.

Updated all files to include the changes.

Added 'GUI.py' file to be the all-encompassing file for Graphical User Interface methods and functions.
The 'clbot_GUI.py' will now become obsolete because it was never truly organized.

Currently working on adding GUI functionality to the different steps of 'Create A Post'. Researching how to connect radio button presses to another function.