Updated 'GUI.py' to include a new class 'Options' that will be the new dialog window for selecting a radio button during the 'create a post' process. Changed the 'Window' class to now be a 'QMainWindow' object that has a "continue" and an "alternative" button. The "continue" button will use the user's inputted values and determine what type of post the user wants to make, then ask the user for required information to create the post. The "alternative" button will take the user to the first step of "create a post" process, "Choose Type of Posting".
Updated 'mech_class.py' with a few minor changes. A minor change in the 'determine_page()' function to print the title of the browser page at the beginning of the function instead of inside the if statements. Another minor change was to the format of calling upon the GUI. Now, the initialize function of 'CreatePost' class start the GUI.